adding support for rest + kafka for reporting, using sink
