+++ metadata-ingestion/tests/integration/dbt/test_dbt.py
@@ -8,7 +8,8 @@ from freezegun import freeze_time
 
 from datahub.configuration.common import DynamicTypedConfig
 from datahub.ingestion.api.ingestion_job_checkpointing_provider_base import JobId
-from datahub.ingestion.run.pipeline import Pipeline, PipelineConfig, SourceConfig
+from datahub.ingestion.run.pipeline import Pipeline
+from datahub.ingestion.run.pipeline_config import PipelineConfig, SourceConfig
 from datahub.ingestion.source.dbt import (
     DBTConfig,
     DBTEntitiesEnabled,
